Moves most fixtures from test_CapData to conftest
